correcao

+++ OsTresMotoqueirosDoApocalipseVerde/OsTresMotoqueirosDoApocalipseVerde/Application/DTOs/Request/CreateEnderecoRequest.cs
@@ -2,7 +2,7 @@
 {
     public class CreateEnderecoRequest
     {
-        public long IdEndereco { get; set; }
+        public int IdEndereco { get; set; }
         public int Numero { get; set; }
         public string Estado { get; set; }
         public string CodigoPais { get; set; }
@@ -10,6 +10,6 @@
         public string Complemento { get; set; }
         public string Rua { get; set; }
 
-        public long FilialId { get; set; }
+        public int FilialId { get; set; }
     }
 }
