+++ OsTresMotoqueirosDoApocalipseVerde/OsTresMotoqueirosDoApocalipseVerde/Application/DTOs/Request/CreateFilialRequest.cs
@@ -2,7 +2,7 @@
 {
     public class CreateFilialRequest
     {
-        public long IdFilial { get; set; }
+        public int IdFilial { get; set; }
         public string NomeFilial { get; set; }
         public string Estado { get; set; }
         public string CodigoPais { get; set; }
@@ -10,6 +10,6 @@
         public string Complemento { get; set; }
         public string Rua { get; set; }
 
-        public long FilialId { get; set; }
+        public int FilialId { get; set; }
     }
 }
