+++ OsTresMotoqueirosDoApocalipseVerde/OsTresMotoqueirosDoApocalipseVerde/Application/DTOs/Response/CreateFilialResponse.cs
@@ -2,7 +2,7 @@
 {
     public class CreateFilialResponse
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string CPF { get; set; }
         public string Telefone { get; set; }
         public string Email { get; set; }
@@ -10,7 +10,7 @@
         public string Senha { get; set; }
         public string Nome { get; set; }
 
-        public long FuncionarioId { get; set; }
-        public long MotoristaId { get; set; }
+        public int FuncionarioId { get; set; }
+        public int MotoristaId { get; set; }
     }
 }
